server/app/requirements.txt
alembic==1.9.2
aniso8601==9.0.1
anyio==3.6.2
asyncio==3.4.3
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.0
fastapi==0.89.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
orjson==3.8.5
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
pyrsistent==0.19.3
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
